crosstool-ng: Take on patch to sort old objects
This patch was introduced in 5cb8bb4acc5a59296306dd8af5837bc1e156763f
to make libstdc++.a reproducible.  But 5cb8bb applied it only to
the host gcc, and the ultimate host /lib64/libstdc++.a comes from
the gcc compiled as part of cross-toolchain.  So we need to apply
it to that gcc also.