checksums: Disregard /usr/src/egg
Rationale: It's actually fine if the egg sources are different, as
long as everything else is the same. Some changes to egg sources are
obviously irrelevant to the goal of reproducibility -- such as updating
the link to download a tarball from (e.g. be8191a), or tweaking the
documentation (e.g., 482dd51) -- but this also includes changes that
just make the build work on a wider variety of systems (e.g., 7c63e6a).

Fixes #12