Download to downloads/, not package dirs
The main rationale is that there's often older versions of sources
lying around after we bump a version, and they're harder to handle
well when the sources are mixed in with the other files in the
package directory.  This way we can easily copy only the sources
actually listed to the result tree.

Also, busybox find doesn't support -execdir, making it inconvenient
to download sources using the old procedure in the README.

We link just the listed sources into the working directory (instead of,
say, providing a $downloads variable pointing to the whole downloads
directory) to make it easier to detect when a build script is using
a source that's not listed in its sources file.